qTcurrent.py : overflow problem solved
The overflow problem was happening due to using SI units for the k_B * T product (Joules) while using eV for all other energy quantities. Therefore, the k-B * T product was close to zero compared in the eV scale used. The problem is solved converting units to eV.

After solving this, the change can be seen in the graph, where the "width" of the part of the Fermi-Dirac function that changes from 1 to zero is about k_B * T (so now, this width is bigger than before solving the overflow problem).